Commit for updates per feedback
Match tickets update table still under work (need to allow seatTicketID update) summary of updates:   TA (Quinn Downey): Formatted dates using DATE_FORMAT() function for clear reading.

Added order date to match tickets table. Schema update to define Matchtickets price attribute and Orders total attribute to decimal(10, 2). CustomerID can no longer be updated. Select statements include listed attributes as opposed to *.

Kurin Schirm: delete button added to match tickets page, this meets criteria for an M:N relationship delete.

Update operation added to matchTickets page, that allows seat reassigning, price change, and matchDate this was update to meet criteria for M:N.

+++ DDL.sql
@@ -54,7 +54,7 @@ CREATE TABLE Orders (
     orderID INT AUTO_INCREMENT PRIMARY KEY,
     customerID INT NOT NULL,
     orderDate DATETIME NOT NULL,
-    total INT NOT NULL,
+    total Decimal(10,2) NOT NULL,
     paymentStatus VARCHAR(10) NOT NULL CHECK (paymentStatus IN ('Paid', 'Processing', 'Cancelled', 'Refunded')),
     FOREIGN KEY (customerID) REFERENCES Customers(customerID)
 );
@@ -72,7 +72,7 @@ CREATE TABLE MatchTickets (
     matchID INT NOT NULL,
     seatID INT NOT NULL,
     orderID INT,
-    price INT NOT NULL,
+    price Decimal(10, 2) NOT NULL,
     FOREIGN KEY (matchID) REFERENCES Matches(matchID) ON DELETE CASCADE,
     FOREIGN KEY (seatID) REFERENCES Seats(seatID) ON DELETE CASCADE,
     FOREIGN KEY (orderID) REFERENCES Orders(orderID) ON DELETE CASCADE
